Time: 10 ms (25.30%) | Memory: 12.9 MB (35.88%) - LeetSync

+++ 64-minimum-path-sum/minimum-path-sum.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
-    long minPathSum(vector<vector<int>>& grid) {
+    int minPathSum(vector<vector<int>>& grid) {
         int m=grid.size(),n=grid[0].size();
-        vector<vector<long>> dp(m,vector<long>(n));
+        vector<vector<int>> dp(m,vector<int>(n));
         dp[0][0]=grid[0][0];
         for(int i=1;i<m;i++) dp[i][0]=dp[i-1][0]+grid[i][0];
         for(int i=1;i<n;i++) dp[0][i]=dp[0][i-1]+grid[0][i];
